+++ test/modules/Auction/MockAtomicAuctionModule.sol
@@ -103,11 +103,11 @@ contract MockAtomicAuctionModule is AuctionModule {
         revert Auction_NotImplemented();
     }
 
-    function claimProceeds(uint96) external pure override returns (uint96, uint96, uint96) {
+    function claimProceeds(uint96) external pure override returns (uint256, uint96, uint96) {
         revert Auction_NotImplemented();
     }
 
-    function _claimProceeds(uint96) internal pure override returns (uint96, uint96, uint96) {
+    function _claimProceeds(uint96) internal pure override returns (uint256, uint96, uint96) {
         revert Auction_NotImplemented();
     }
